[FUN] Ensemble Cobinet
Externalize Correlation Ensemble into Dedicated Module
